cortex/requirements.txt
fastapi
uvicorn
websockets
python-multipart
pydantic
email-validator
supabase
python-dotenv
google-generativeai
google-genai
httpx
